Fix a timing problem caused by improper use of asynchCommand().
If you use asynchCommand() to call an action command in a subsystem, you should call get() on the Future object it returns before sending the next action command, or you should use synchCommand() instead. This way the subsystem will be ready for another action command. Here we were just throwing the Future away.